+++ Toggly.FeatureManagement.Javascript/feature_flags_toggly/spec/websocket.spec.ts
@@ -115,7 +115,7 @@ describe('Toggly WebSocket', () => {
       });
 
       expect(latestWs()).toBeDefined();
-      expect(latestWs().url).toBe('wss://definitions.toggly.io/my-app-key/ws?sdk=javascript&sdkVersion=1.3.1');
+      expect(latestWs().url).toBe('wss://definitions.toggly.io/my-app-key/ws?sdk=javascript&sdkVersion=1.9.0');
     });
 
     it('sets wsConnected to true on open', async () => {